scraper.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict
import random
from datetime import datetime
import time

class NewsArticleScraper:
    def __init__(self):
        """Initialize the news scraper with common headers and sources."""
        self.user_agents = [
            'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
            'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:89.0) Gecko/20100101 Firefox/89.0',
            'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.1.1 Safari/605.1.15'
        ]
        
        self.sources = {
            'techcrunch': {
                'base_url': 'https://techcrunch.com',
                'article_selector': 'article.post-block',
                'title_selector': 'h2.post-block__title a',
                'content_selector': 'div.article-content'
            },
            'theverge': {
                'base_url': 'https://www.theverge.com/tech',
                'article_selector': 'h2.c-entry-box--compact__title',
                'title_selector': 'a',
                'content_selector': 'div.c-entry-content'
            },
            'wired': {
                'base_url': 'https://www.wired.com/tag/technology',
                'article_selector': 'div.SummaryItemContent-gYA-Dbz',
                'title_selector': 'h3 a',
                'content_selector': 'div.body__content'
            }
        }

    def _get_random_headers(self) -> Dict:
        """Generate random headers to avoid detection."""
        return {
            'User-Agent': random.choice(self.user_agents),
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
            'Connection': 'keep-alive',
        }

    def _clean_text(self, text: str) -> str:
        """Clean and normalize text content."""
        if not text:
            return ""
        # Remove extra whitespace
        text = ' '.join(text.split())
        # Remove special characters
        text = text.replace('\n', ' ').replace('\t', ' ').strip()
        return text

    def _extract_article_content(self, url: str, content_selector: str) -> str:
        """Extract the main content from an article URL."""
        try:
            response = requests.get(url, headers=self._get_random_headers(), timeout=10)
            response.raise_for_status()
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Try the specific content selector first
            content = soup.select_one(content_selector)
            if not content:
                # Fallback to common content containers
                content = soup.select_one('article') or \
                         soup.select_one('main') or \
                         soup.select_one('.article-body') or \
                         soup.select_one('.post-content')
            
            if content:
                # Get all paragraphs
                paragraphs = content.find_all('p')
                text = ' '.join(p.get_text() for p in paragraphs)
                return self._clean_text(text)
            
            return ""
        except Exception as e:
            return ""

    def scrape_articles(self, limit: int = 5) -> List[Dict]:
        """Scrape articles from configured sources."""
        articles = []
        
        for source_name, source_config in self.sources.items():
            try:
                response = requests.get(
                    source_config['base_url'],
                    headers=self._get_random_headers(),
                    timeout=10
                )
                response.raise_for_status()
                soup = BeautifulSoup(response.text, 'html.parser')
                
                # Find all article elements
                article_elements = soup.select(source_config['article_selector'])
                
                for article in article_elements[:limit]:
                    try:
                        # Extract title and URL
                        title_element = article.select_one(source_config['title_selector'])
                        if not title_element:
                            continue
                            
                        title = self._clean_text(title_element.get_text())
                        url = title_element.get('href', '')
                        
                        # Handle relative URLs
                        if url.startswith('/'):
                            url = source_config['base_url'] + url
                        elif not url.startswith('http'):
                            continue
                        
                        # Extract content
                        content = self._extract_article_content(url, source_config['content_selector'])
                        
                        # Validate article
                        if len(content) > 500 and title and url:  # Minimum content length
                            articles.append({
                                'title': title,
                                'url': url,
                                'content': content,
                                'timestamp': datetime.now().isoformat(),
                                'source': source_name
                            })
                            
                            if len(articles) >= limit:
                                return articles
                            
                            # Be nice to the servers
                            time.sleep(2)
                    
                    except Exception as e:
                        continue
                        
            except Exception as e:
                continue
        
        # If no articles were found, add a sample article
        if not articles:
            articles.append({
                'title': 'Sample Technology Article',
                'url': 'https://example.com/sample',
                'content': 'This is a sample technology article that discusses artificial intelligence, ' \
                          'machine learning, and their impact on society. AI has revolutionized many industries ' \
                          'including healthcare, finance, and transportation. Machine learning algorithms continue ' \
                          'to improve, making systems more efficient and capable of handling complex tasks.',
                'timestamp': datetime.now().isoformat(),
                'source': 'sample'
            })
        
        return articles
