requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
streamlit>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
newspaper3k>=0.2.8
scikit-learn>=1.6.1
faiss-cpu>=1.7.4
